Refactor the optimisation tests.
The platypus and pygmo tests are in separate modules. These are skipped if the respective
package can not be imported.
